requirements.txt
numpy==2.4.2
onnx==1.20.1
onnxruntime==1.24.2
pillow==12.1.1